requirements.txt
opencv-contrib-python
Pillow
torch
torchvision
opencv-python
boto3
botocore
